Added Calculate.cs file and GenerateSecretNimber

+++ GuessingGameApi/GuessingGameApi/Entities/Game.cs
@@ -7,9 +7,9 @@ public class Game
 {
     [Key,DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public long Id { get; set; }
-    public ushort SecretNumber { get; set; }
-    public ushort NumberOfTries { get; set; }
-    public ushort MaximumTries { get; set; }
+    public int SecretNumber { get; set; }
+    public int NumberOfTries { get; set; }
+    public int MaximumTries { get; set; }
     public bool IsFinish { get; set; }
     public bool IsWinner { get; set; }
     public long UserId { get; set; }
